+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixes
 
 - [rc.2] Fix an issue where the scroll edge effect is not applied to the file browser.
+- [rc.2] Fix an issue where a input field in the print dialog is compressed.