+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Fix an issue where the “Convert Indentation” commands don’t convert if no characters are selected.
 - Fix an issue where the input fields in the Find window accept rich text.
+- Address a potential crash on macOS 26 that occurs when performing a keyboard shortcut.
 - [trivial] Fix an issue where the default encoding pop-up button in the Format settings pane is positioned incorrectly.