+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix an issue where the file browser doesn't show the entire filenames when the scroller slot is shown.
 - Fix an issue where the “Convert Indentation” commands don’t convert if no characters are selected.
 - Fix an issue where the input fields in the Find window accept rich text.
+- Fix an issue where folder documents can open multiple windows for the same folder.
 - Address a potential crash on macOS 26 that occurs when performing a keyboard shortcut.
 - Address a potential crash on macOS 26 that can occur when opening a document on an Intel-based Mac.
 - [trivial] Fix an issue where the default encoding pop-up button in the Format settings pane is positioned incorrectly.